+++ connect-template.php
@@ -31,4 +31,6 @@
     $userpass = "";
 
     $bsa_benodigd = 42;
+    $afbeelding_dir = "graphs";
+    $grafiek = True;
 ?>